+++ buildpacks/php/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2025-08-14
+
 ### Changed
 
 - Retry downloads during bootstrapping ([#226](https://github.com/heroku/buildpacks-php/pull/226))
@@ -141,7 +143,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation
 
-[unreleased]: https://github.com/heroku/buildpacks-php/compare/v1.0.6...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-php/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/heroku/buildpacks-php/compare/v1.0.6...v1.1.0
 [1.0.6]: https://github.com/heroku/buildpacks-php/compare/v1.0.5...v1.0.6
 [1.0.5]: https://github.com/heroku/buildpacks-php/compare/v1.0.4...v1.0.5
 [1.0.4]: https://github.com/heroku/buildpacks-php/compare/v1.0.3...v1.0.4